Conflicts:
	src/Framework/Module/Administration/View/form/textarea.html.twig

+++ src/Framework/Module/HttpServer/HttpServer.php
@@ -15,6 +15,7 @@ use Framework\Http\Response as KernelResponse;
 use Framework\Module\EventDispatcher\EventDispatcher;
 use React\EventLoop\Factory;
 use React\Filesystem\Filesystem;
+use React\Http\Handler\RequestHandler;
 use React\Http\Processor\FormField;
 use React\Socket\Server;
 
@@ -36,11 +37,6 @@ class HttpServer extends Service
     protected $loop;
 
     /**
-     * @var Filesystem
-     */
-    protected $filesystem;
-
-    /**
      * @throws \React\Socket\ConnectionException
      */
     public function initialize()
@@ -48,8 +44,6 @@ class HttpServer extends Service
         $this->loop = Factory::create();
         $this->socket = new Server($this->loop);
         $this->http = new \React\Http\Server($this->socket);
-
-        $this->filesystem = Filesystem::create($this->loop);
 
         /** @var EventDispatcher $dispatcher */
         $dispatcher = $this->container->get('event_dispatcher');
@@ -60,57 +54,50 @@ class HttpServer extends Service
                 $request->getMethod(),
                 $request->getPath(),
                 $request->getQuery(),
-                $request->getHeaders(),
+                $request->headers,
                 $request->getVersion()
             );
-
-            $request->on('form.file', function (FormField $field, $originalFilename) use ($kernelRequest) {
-
-                $file = new File();
-                $kernelRequest->files->set($field->getName(), $file);
-
-                $filename = tempnam(sys_get_temp_dir(), '');
-
-                $file->setName($filename);
-                $file->setOriginalName($originalFilename);
-
-//                $fileDescriptor = $this->filesystem->file($filename);
-
-                $field->on('data', function ($data) use ($filename) {
-                    file_put_contents($filename, $data, FILE_APPEND | FILE_BINARY);
-//                    $fileDescriptor->open('cw')->then(function ($stream) use ($data) {
-//                        $stream->write($data);
-//                        $stream->end();
-//                    });
-                });
-
-                $field->on('end', function ($data) use ($filename) {
-                    file_put_contents($filename, $data, FILE_APPEND);
-//                    $fileDescriptor->open('cw')->then(function ($stream) use ($data) {
-//                        $stream->write($data);
-//                        $stream->end();
-//                    });
-                });
-            });
-
-            $request->on('form.field', function (FormField $field) use ($kernelRequest) {
-                $field->on('data', function ($data) use ($kernelRequest, $field) {
-                    $kernelRequest->atributes->set($field->getName(), $data);
-                });
-            });
-
-            $request->on('end', function () use ($kernelRequest) {
-                $kernelRequest->emit('ready');
-            });
 
             $kernelRequest->on('response', function (KernelResponse $kernelResponse) use ($response) {
                 $response->writeHead($kernelResponse->getStatusCode(), $kernelResponse->getHeaders());
                 $response->end($kernelResponse->getBody());
             });
 
-            $kernelRequest->on('ready', function () use ($dispatcher, $kernelRequest) {
+            $kernelRequest->on('end', function () use ($dispatcher, $kernelRequest) {
                 $dispatcher->dispatch('request', [$kernelRequest]);
             });
+
+            $requestHandler = new RequestHandler();
+
+            $requestHandler->on('end', function () use ($kernelRequest) {
+                $kernelRequest->emit('end');
+            });
+
+            $requestHandler->on('data', function (FormField $field) use ($kernelRequest) {
+                if ($field->isFile()) {
+
+                    $filename = tempnam(sys_get_temp_dir(), '');
+
+                    $dataHandler = function ($data) use ($filename, $kernelRequest, $field) {
+                        if ($kernelRequest->files->get($field->getName()) === null) {
+                            $file = new File();
+                            $file->setName($filename);
+                            $file->setOriginalName($field->attributes->get(FormField::ORIGINAL_FILENAME));
+                            $kernelRequest->files->set($field->getName(), $file);
+                        }
+
+                        file_put_contents($filename, $data, FILE_APPEND | FILE_BINARY);
+                    };
+                } else {
+                    $dataHandler = function ($data) use ($kernelRequest, $field) {
+                        $kernelRequest->atributes->set($field->getName(), $data);
+                    };
+                }
+
+                $field->on('data', $dataHandler);
+            });
+
+            $requestHandler->handle($request);
         });
 
         $this->socket->listen(8080, '0.0.0.0');
